emplacement projet aradon

+++ projetAradon/code/src/Controllers/UserController.php
@@ -44,9 +44,10 @@ class UserController
         }
 
         $errors = array();
+        var_dump($_SESSION);
         if (!isset($_POST) || !empty($_POST)) {
-            if (empty($_POST['username']) || !preg_match('/^[a-zA-Z0-9_]+$/', $_POST['username'])) {
-                $errors['username'] = "votre pseudo n'est pas valide (alphanumérique)";
+            if (empty($_POST['pseudo']) || !preg_match('/^[a-zA-Z0-9_]+$/', $_POST['pseudo'])) {
+                $errors['pseudo'] = "votre pseudo n'est pas valide (alphanumérique)";
             }
 
             if (empty($_POST['email']) || !filter_var($_POST['email'], FILTER_VALIDATE_EMAIL)) {
@@ -56,24 +57,26 @@ class UserController
             if (empty($_POST['dateOfBirth']) || !strtotime($_POST['dateOfBirth'])) {
                 $errors['dateOfBirth'] = "votre date de naissance est invalide";
             }
-            if (empty($_POST['adresse'])) {
-                $errors['adresse'] = "votre adresse est invalide";
-            }
-            $newAvatar = $_FILES['profilPicture'];
-            $newFileNameAvatar = Functions::getRandomiseImageName($newAvatar['name']);
-            Functions::imageValidation($newAvatar, $newFileNameAvatar, ENV::$AVATAR_PATH);
-            if ($_SESSION['currentUser']->getProfilPicture()) {
-
-                unlink(Env::$AVATAR_PATH . $_SESSION['currentUser']->getProfilPicture());
+            if (empty($_POST['address'])) {
+                $errors['address'] = "votre adresse est invalide";
             }
 
+            if ($_FILES['profilPicture']['size']) {
+                $newAvatar = $_FILES['profilPicture'];
+                $newFileNameAvatar = Functions::getRandomiseImageName($newAvatar['name']);
+                Functions::imageValidation($newAvatar, $newFileNameAvatar, ENV::$AVATAR_PATH);
+                if ($_SESSION['currentUser']->getProfilPicture()) {
+                    unlink(Env::$AVATAR_PATH . $_SESSION['currentUser']->getProfilPicture());
+                }
+            }
+    
             $updatedUser = array(
                 'id' => $_SESSION['currentUser']->getId(),
-                'pseudo' => $_POST['username'],
+                'pseudo' => $_POST['pseudo'],
                 'email' => $_POST['email'],
-                'adresse' => $_POST['adresse'],
+                'address' => $_POST['address'],
                 'dateOfBirth' => $_POST['dateOfBirth'],
-                'profilPicture' => $newFileNameAvatar
+                'profilPicture' => $newFileNameAvatar ?? $_SESSION['currentUser']->getProfilPicture()
             );
 
             $this->userRepository->update($updatedUser);
